CIV-9198 - renamed two file camelCase and added test for hearing duration formatter
