+++ src/main/common/models/caseProgression/hearingDuration.ts
