+++ src/main/routes/features/caseProgression/trialArrangements/hearingDurationAndOtherInformationController.ts
@@ -11,7 +11,7 @@ import {caseNumberPrettify} from 'common/utils/stringUtils';
 import {constructResponseUrlWithIdParams} from 'common/utils/urlFormatter';
 import {GenericForm} from 'form/models/genericForm';
 
-import {OtherTrialInformation} from 'form/models/caseProgression/trialArrangements/OtherTrialInformation';
+import {OtherTrialInformation} from 'form/models/caseProgression/trialArrangements/otherTrialInformation';
 import {saveCaseProgression} from 'services/features/caseProgression/caseProgressionService';
 import {Claim} from 'models/claim';
 import {getClaimById} from 'modules/utilityService';
